backend/src/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.110.1
h11==0.14.0
idna==3.10
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
ruff==0.6.4
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.29.0
